+++ src/types/plugin.ts
@@ -1,0 +1,3 @@
+import {App} from 'vue';
+
+export type UserPlugin = (ctx: {app: App}) => void;